+++ src/event_imp.rs
@@ -280,22 +280,6 @@ impl PollOpt {
         PollOpt::ONESHOT
     }
 
-    #[deprecated(since = "0.6.5", note = "removed")]
-    #[cfg(feature = "with-deprecated")]
-    #[doc(hidden)]
-    #[inline]
-    pub fn urgent() -> PollOpt {
-        PollOpt(0b1000)
-    }
-
-    #[deprecated(since = "0.6.5", note = "removed")]
-    #[cfg(feature = "with-deprecated")]
-    #[doc(hidden)]
-    #[inline]
-    pub fn all() -> PollOpt {
-        PollOpt::EDGE | PollOpt::LEVEL | PollOpt::ONESHOT
-    }
-
     /// Returns true if the options include edge-triggered notifications.
     ///
     /// See [`Poll`] for more documentation on polling.
@@ -354,23 +338,6 @@ impl PollOpt {
     #[inline]
     pub fn is_oneshot(&self) -> bool {
         self.contains(PollOpt::ONESHOT)
-    }
-
-    #[deprecated(since = "0.6.5", note = "removed")]
-    #[cfg(feature = "with-deprecated")]
-    #[doc(hidden)]
-    #[allow(deprecated)]
-    #[inline]
-    pub fn is_urgent(&self) -> bool {
-        self.contains(PollOpt::urgent())
-    }
-
-    #[deprecated(since = "0.6.5", note = "removed")]
-    #[cfg(feature = "with-deprecated")]
-    #[doc(hidden)]
-    #[inline]
-    pub fn bits(&self) -> usize {
-        self.0
     }
 
     /// Returns true if `self` is a superset of `other`.
@@ -492,18 +459,6 @@ impl ops::Sub for PollOpt {
     }
 }
 
-#[deprecated(since = "0.6.10", note = "removed")]
-#[cfg(feature = "with-deprecated")]
-#[doc(hidden)]
-impl ops::Not for PollOpt {
-    type Output = PollOpt;
-
-    #[inline]
-    fn not(self) -> PollOpt {
-        PollOpt(!self.0)
-    }
-}
-
 impl fmt::Debug for PollOpt {
     fn fmt(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
         let mut one = false;
@@ -594,20 +549,6 @@ impl Ready {
     /// [`Poll`]: struct.Poll.html
     pub const EMPTY: Ready = Ready(0);
 
-    #[deprecated(since = "0.6.5", note = "use Ready::EMPTY instead")]
-    #[cfg(feature = "with-deprecated")]
-    #[doc(hidden)]
-    pub fn none() -> Ready {
-        Ready::EMPTY
-    }
-
-    #[deprecated(since = "0.6.10", note = "use Ready::EMPTY instead")]
-    #[cfg(feature = "with-deprecated")]
-    #[doc(hidden)]
-    pub fn empty() -> Ready {
-        Ready::EMPTY
-    }
-
     /// Returns a `Ready` representing readable readiness.
     ///
     /// See [`Poll`] for more documentation on polling.
@@ -658,31 +599,6 @@ impl Ready {
         Ready::WRITABLE
     }
 
-    #[deprecated(since = "0.6.5", note = "use UnixReady instead")]
-    #[cfg(feature = "with-deprecated")]
-    #[doc(hidden)]
-    #[inline]
-    pub fn error() -> Ready {
-        Ready(ERROR)
-    }
-
-    #[deprecated(since = "0.6.5", note = "use UnixReady instead")]
-    #[cfg(feature = "with-deprecated")]
-    #[doc(hidden)]
-    #[inline]
-    pub fn hup() -> Ready {
-        Ready(HUP)
-    }
-
-    #[deprecated(since = "0.6.5", note = "removed")]
-    #[cfg(feature = "with-deprecated")]
-    #[doc(hidden)]
-    #[inline]
-    pub fn all() -> Ready {
-        Ready::READABLE |
-            Ready::WRITABLE
-    }
-
     /// Returns true if `Ready` is the empty set
     ///
     /// See [`Poll`] for more documentation on polling.
@@ -698,14 +614,6 @@ impl Ready {
     #[inline]
     pub fn is_empty(&self) -> bool {
         *self == Ready::EMPTY
-    }
-
-    #[deprecated(since = "0.6.5", note = "use Ready::is_empty instead")]
-    #[cfg(feature = "with-deprecated")]
-    #[doc(hidden)]
-    #[inline]
-    pub fn is_none(&self) -> bool {
-        self.is_empty()
     }
 
     /// Returns true if the value includes readable readiness
@@ -748,22 +656,6 @@ impl Ready {
         self.contains(Ready::WRITABLE)
     }
 
-    #[deprecated(since = "0.6.5", note = "use UnixReady instead")]
-    #[cfg(feature = "with-deprecated")]
-    #[doc(hidden)]
-    #[inline]
-    pub fn is_error(&self) -> bool {
-        self.contains(Ready(ERROR))
-    }
-
-    #[deprecated(since = "0.6.5", note = "use UnixReady instead")]
-    #[cfg(feature = "with-deprecated")]
-    #[doc(hidden)]
-    #[inline]
-    pub fn is_hup(&self) -> bool {
-        self.contains(Ready(HUP))
-    }
-
     /// Adds all readiness represented by `other` into `self`.
     ///
     /// This is equivalent to `*self = *self | other`.
@@ -802,14 +694,6 @@ impl Ready {
     pub fn remove<T: Into<Self>>(&mut self, other: T) {
         let other = other.into();
         self.0 &= !other.0;
-    }
-
-    #[deprecated(since = "0.6.5", note = "removed")]
-    #[cfg(feature = "with-deprecated")]
-    #[doc(hidden)]
-    #[inline]
-    pub fn bits(&self) -> usize {
-        self.0
     }
 
     /// Returns true if `self` is a superset of `other`.
@@ -921,18 +805,6 @@ impl<T: Into<Ready>> ops::SubAssign<T> for Ready {
     }
 }
 
-#[deprecated(since = "0.6.10", note = "removed")]
-#[cfg(feature = "with-deprecated")]
-#[doc(hidden)]
-impl ops::Not for Ready {
-    type Output = Ready;
-
-    #[inline]
-    fn not(self) -> Ready {
-        Ready(!self.0)
-    }
-}
-
 impl fmt::Debug for Ready {
     fn fmt(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
         let mut one = false;
@@ -1029,13 +901,6 @@ impl Event {
     /// assert_eq!(event.readiness(), Ready::READABLE);
     /// ```
     pub fn readiness(&self) -> Ready {
-        self.kind
-    }
-
-    #[deprecated(since = "0.6.5", note = "use Event::readiness()")]
-    #[cfg(feature = "with-deprecated")]
-    #[doc(hidden)]
-    pub fn kind(&self) -> Ready {
         self.kind
     }